Add dnd4data.js, round out dnd4model.js
This also modifies cbimporter to do the right thing with regard to
prefixes and such. Now we should be able to copy/paste JS from the
preview or batch conversion right into dnd4data.js.

We are also taking this moment to stub out the new properties in
dnd4model.js. This also serves as an example of how to extend the
model in model.js: these properties are specific to dnd 4e.