Some UI enhancements, selection support for choices.
Added a hacky UI for race in addition to class, as a byproduct,
abstracted out support for a master/detail kind of view. In addition,
made it so that selecting the item in the list changed the selection
in the engine. That meant adding support for selection in the engine,
leading to new tests.