Detail for an element can now use an iframe
It turns out that you get GREAT descriptions from the D&D insider
compendium; so we support that now, via an optional member of the
element that provides a "compendium URL." The converter fills it out
for entries that it knows the compendium has. Other folks can put
custom URLs there. Otherwise we need a default. Hmmm....