Class Feature Selection; Fix Transitive Grants
There was a nasty bug where we were doing reference counting twice;
this caused transitive grants to get to -1 references by the time we
were trying to process their undo rules. We found it by trying to
build a little piece of UI for class feature selection.

Now the UI is in and the fix is in and there's a test. Oh good!