Converter: Don't emit specifics
Specifics are used for two things: to render the pretty HTML and
sometimes to extract data. We're going to render the pretty HTML
offline, and so all we need is the data. But we don't know which data
yet. So we're cutting out specifics entirely and we'll re-add the
relevant data later.